refactor(build-studio): keep engine modules focused
